requirements.txt
# -------------------------------
# Core FastAPI Backend
# -------------------------------
fastapi==0.115.0
uvicorn[standard]==0.32.0

# -------------------------------
# Authentication + Security
# -------------------------------
python-jose==3.3.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4

# -------------------------------
# Database
# -------------------------------
pymysql==1.1.1

# -------------------------------
# CORS and Utils
# -------------------------------
python-dotenv==1.0.1
requests==2.32.3

# -------------------------------
# Optional (for Django hash compatibility, if you import Django)
# -------------------------------
django==5.1.2

# -------------------------------
# Data and Debugging
# -------------------------------
pandas==2.2.3
numpy==1.26.4
